+++ idf_component_manager/idf_extensions.py
@@ -4,7 +4,7 @@
 import sys
 import warnings
 
-from idf_component_manager.utils import error, warn
+from idf_component_manager.utils import error, showwarning
 from idf_component_tools.errors import FatalError
 
 from .core import ComponentManager
@@ -62,11 +62,9 @@ VERSION_PARAMETER = [
 def action_extensions(base_actions, project_path):
     def callback(subcommand_name, ctx, args, **kwargs):
         try:
-            with warnings.catch_warnings(record=True) as w:
-                manager = ComponentManager(args.project_dir)
-                getattr(manager, str(subcommand_name).replace('-', '_'))(kwargs)
-                for warning in w:
-                    warn(warning.message)
+            warnings.showwarning = showwarning
+            manager = ComponentManager(args.project_dir)
+            getattr(manager, str(subcommand_name).replace('-', '_'))(kwargs)
         except FatalError as e:
             error(e)
             sys.exit(e.exit_code)